Feat: swagger API에 태그 및 오퍼레이션 추가
