+++ src/main/java/kr/hhplus/be/server/interfaces/api/user/UserController.java
@@ -1,18 +1,24 @@
 package kr.hhplus.be.server.interfaces.api.user;
 
+import io.swagger.v3.oas.annotations.Operation;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
 import kr.hhplus.be.server.common.response.ApiResponse;
+import kr.hhplus.be.server.common.response.ErrorResponse;
 import kr.hhplus.be.server.common.response.ResponseCode;
 import kr.hhplus.be.server.interfaces.api.dto.UserBalanceRequest;
 import kr.hhplus.be.server.interfaces.api.dto.UserBalanceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-
+@Tag(name = "잔액 충전/조회 API", description = "잔액을 충전하고 조회하는 api 입니다.")
 @RestController
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
 public class UserController {
 
     // 잔액 조회 API
+    @Operation(summary = "잔액 조회")
     @GetMapping("/balance")
     public ApiResponse<UserBalanceResponse> getBalance(){
         UserBalanceResponse userBalanceResponse = new UserBalanceResponse(12345L, 100.00);
@@ -20,6 +26,7 @@ public class UserController {
     }
 
     // 잔액 충전 API
+    @Operation(summary = "잔액 충전")
     @PostMapping("/balance/charge")
     public ApiResponse<UserBalanceResponse> chargeBalance(@RequestBody UserBalanceRequest userBalanceRequest){
